refactor: integrate node_api
